change RideFareCalculationStrategy signature from double to BigDecimal

+++ src/main/java/com/akhil/uber_backend/uber_ride/strategies/RideFareCalculationStrategy.java
@@ -3,10 +3,12 @@ package com.akhil.uber_backend.uber_ride.strategies;
 import com.akhil.uber_backend.uber_ride.dto.RideRequestDTO;
 import com.akhil.uber_backend.uber_ride.models.RideRequest;
 
+import java.math.BigDecimal;
+
 public interface RideFareCalculationStrategy {
 
     double RIDE_FARE_MULTIPLIER = 10;
 
-    double calculateFare(RideRequest rideRequest);
+    BigDecimal calculateFare(RideRequest rideRequest);
 
 }
